Reorganize code in source files for error logging/misc functions
