+++ compiler/include/misc.h
@@ -38,6 +38,14 @@
 #define exit(x) dont_use_exit_use_clean_exit_instead
 #endif
 
+#ifndef NO_RETURN
+  #if defined(__GNUC__) && __GNUC__ >= 3
+    #define NO_RETURN __attribute__((__noreturn__))
+  #else
+    #define NO_RETURN
+  #endif
+#endif
+
 #define iterKindTypename          "iterKind"
 #define iterKindLeaderTagname     "leader"
 #define iterKindFollowerTagname   "follower"
@@ -53,13 +61,31 @@ enum class GpuCodegenType {
   GPU_CG_AMD_HIP,
 };
 
-bool        forceWidePtrsForLocal();
-bool        requireWideReferences();
-bool        requireOutlinedOn();
-bool        usingGpuLocaleModel();
 GpuCodegenType getGpuCodegenType();
 
+bool forceWidePtrsForLocal();
+bool requireWideReferences();
+
+//
+// If the --no-local flag is used, or the locale model is not 'flat'
+// (i.e., has sub-locales that an on-clause might target), we should
+// require on-clauses to be "outlined" (i.e., we should not assume the
+// on-clause is a no-op and execute the associated statement locally.
+//
+bool requireOutlinedOn();
+
+// Return true if the current locale model needs GPU code generation.
+bool usingGpuLocaleModel();
+
 const char* cleanFilename(const BaseAST* ast);
-const char* cleanFilename(const char*    name);
+const char* cleanFilename(const char* name);
+
+void startCatchingSignals();
+void stopCatchingSignals();
+
+void clean_exit(int status) NO_RETURN;
+
+// For debugging purposes - print the contents of the static callstack.
+void printCallStackCalls();
 
 #endif
